support docker for running in a defined environment
For testing and debugging it is very helpful to run in a well defined
environment.

This adds a Dockerfile that is based on debian:buster and installs
all required packages to run meta-eid.
It also clones the needed git repositories and does the initial
sbuild setup. It adds the user 'eid'.

Because docker needs some arguments, e.g. for allowing sbuild to
do mounts, a Makefile is added.

A .gitignore file is introduced to ignore the stamp file needed
by make.

The usage is explained in the README.md.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>