enable tests in travis
On github travis can be used for testing.

The .travis.yml file describes the build and test workflow.
Basicly it just enables docker and runs 'make test'.

The test target in the Makefile depends on building the docker
container. Currently it just calls a test script.

The test script executes some simple tests inside the docker
container that are also mentioned in README.md

Signed-off-by: Manuel Traut <manut@linutronix.de>
Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>