docker/test: use ready to use image from dockerhub
Travis runs currently consume about 6 minutes for building the
docker image and the sbuild chroot.

To avoid this, the Dockerfile his hosted in another git repo.
Currently http://github.com/manut/docker-eid-base is used and
connected with dockerhub to automatically build and provide the
image.

An new Dockerfile is added to meta-eid that pulls the prebuild
base image from dockerhub and copies the current version of
meta-eid into the image.

To ensure that the current code is copied into the image before
the testing, the test make target needs to depend on the clean
target.

Signed-off-by: Manuel Traut <manut@linutronix.de>
Signed-off-by: Kazuhiro Hayashi <kazuhiro3.hayashi@toshiba.co.jp>